Use namespace+constexpr for driver constants instead of enum class
enum class requires static_cast at every Port::write() call site in
freestanding C++ (no <type_traits> for template workarounds). The
namespace+constexpr pattern gives the same scoping (AtaCommand::Read,
AtaStatus::Bsy) with implicit conversion to u8, and works uniformly
for both discrete commands and bitwise flags.

+++ src/drivers/keyboard.cpp
@@ -153,7 +153,7 @@ KeyboardCommandByte KeyboardDriver::readCommandByte() {
     KeyboardCommandByte status;
 
     // Requests current command byte.
-    cmd.write(static_cast<u8>(KeyboardCommand::ReadCommandByte));
+    cmd.write(KeyboardCommand::ReadCommandByte);
 
     status.byte = data.read();
     return status;
@@ -185,7 +185,7 @@ void KeyboardDriver::activate() {
     }
 
     // Enable communication with the keyboard.
-    cmd.write(static_cast<u8>(KeyboardCommand::EnableKeyboardInterface));
+    cmd.write(KeyboardCommand::EnableKeyboardInterface);
 
     // Read and modify current command byte.
     KeyboardCommandByte status = readCommandByte();
@@ -193,7 +193,7 @@ void KeyboardDriver::activate() {
     status.disable_keyboard = false;
 
     // Set modified command byte to be the new one.
-    cmd.write(static_cast<u8>(KeyboardCommand::WriteCommandByte));
+    cmd.write(KeyboardCommand::WriteCommandByte);
     data.write(status.byte);
 
     // Enables keyboard and reads the ACK response so it does not remain
